requirements.txt
appdirs==1.4.3
attrs==16.3.0
packaging==16.8
Parsley==1.3
pyparsing==2.2.0
six==1.10.0
